requirements.txt
fastapi
uvicorn
gspread
google-auth
pandas